feat: add ambassador verify endpoint
