+++ packages/espressocash_api/lib/espressocash_api.dart
@@ -1,8 +1,7 @@
 export 'package:jupiter_aggregator/jupiter_aggregator.dart';
 
 export 'src/client.dart';
-export 'src/dto/ambassador_referral.dart';
-export 'src/dto/ambassador_stats.dart';
+export 'src/dto/ambassador.dart';
 export 'src/dto/create_payment.dart';
 export 'src/dto/dln_payment.dart';
 export 'src/dto/durable_transactions.dart';
