--- packages/espressocash_api/lib/src/dto/ambassador_referral.dart
@@ -1,14 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'ambassador_referral.freezed.dart';
-part 'ambassador_referral.g.dart';
-
-@freezed
-class AmbassadorReferralRequestDto with _$AmbassadorReferralRequestDto {
-  const factory AmbassadorReferralRequestDto({
-    required String ambassadorAddress,
-  }) = _AmbassadorReferralRequestDto;
-
-  factory AmbassadorReferralRequestDto.fromJson(Map<String, dynamic> json) =>
-      _$AmbassadorReferralRequestDtoFromJson(json);
-}
